Revert "refactor: use ES modules for audio and storage"
