--- icy-tower/js/store.js
@@ -1,12 +1,0 @@
-export function getLocal(key, defaultValue) {
-  try {
-    const value = localStorage.getItem(key);
-    return value !== null ? JSON.parse(value) : defaultValue;
-  } catch (e) {
-    return defaultValue;
-  }
-}
-
-export function setLocal(key, value) {
-  localStorage.setItem(key, JSON.stringify(value));
-}
